SpringFrameworkSourceCode/src/main/java/com/spring/source/code/beans/ch5/CarFactoryBean.java
package com.spring.source.code.beans.ch5;

import lombok.Getter;
import lombok.Setter;
import org.springframework.beans.factory.FactoryBean;

/**
 * 自定义bean的实现方式
 * @author code
 * @Title: CarFactoryBean
 * @Copyright: Copyright (c) 2017
 * @Description: <br>
 * @Company: www.qdingnet.com
 * @Created on 2018/9/13上午11:24
 */
public class CarFactoryBean implements FactoryBean<Car> {
    @Setter
    @Getter
    private String carInfo;

    @Override
    public Car getObject() throws Exception {
        Car car = new Car();
        String[] infos = carInfo.split(",");
        car.setBrand(infos[0]);
        car.setMaxSpeed(Integer.valueOf(infos[1]));
        car.setPrice(Double.valueOf(infos[2]));

        return car;
    }

    @Override
    public Class<Car> getObjectType() {
        return Car.class;
    }

    @Override
    public boolean isSingleton() {
        return false;
    }


}
